relic: clean API for easier external lib usage
